python/requirements.txt
flask
pyserial
opencv-python-headless
